+++ test/integration/multi_index_test.rb
